Moved Search Conversion tracking back into _app to ensure all pages are handles
